Made config.json a listener in case it changes
So upon first initalization of the app, it would import the config file,
so if it changed it wouldn't update. However with fs.watch we can detect
if it changed. I don't like this as it is not covered in 100% but I
guess it'll do for now. Perhaps there is a package that we can use ?