Added a very very simple electron desktop client to run server.
If I can get this to work it be great because I won't have to rely on
the updater that I created. Which means I wasted a week :^).

To run this you need to have an electron, there is no build for the
desktop client yet. As I have to look more into the building process and
if I have to restructure the project or not.

For references on what I used, see:
https://github.com/frankhale/electron-with-express